Fix stupid error, which happens only on Windows. And we thought this has been fixed. Now my_thread_end() should be called and there should be no warnings in the error logs of exiting thread, which hasn't deinited, and leaks.
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/branches/PHP_5_2@242438 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/mysqli/php_mysqli.h
@@ -241,9 +241,10 @@ PHP_MYSQLI_EXPORT(zend_object_value) mysqli_objects_new(zend_class_entry * TSRML
 #define MYSQLI_RETURN_LONG_LONG(__val) \
 { \
 	if ((__val) < LONG_MAX) {		\
-		RETURN_LONG((__val));		\
+		RETURN_LONG((long) (__val));		\
 	} else {				\
 		char *ret;			\
+		/* always used with my_ulonglong -> %llu */ \
 		int l = spprintf(&ret, 0, "%llu", (__val));	\
 		RETURN_STRINGL(ret, l, 0);		\
 	}					\
